trying to get date selection to work
