+++ OptimalHealth/Models/Database/DailyLogModel.xcdatamodeld/DailyLogModel.xcdatamodel/contents
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="20086" systemVersion="21G115" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
+<model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="21512" systemVersion="22A400" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
     <entity name="DailyLog" representedClassName="DailyLog" syncable="YES" codeGenerationType="class">
-        <attribute name="date" attributeType="Date" usesScalarValueType="NO"/>
+        <attribute name="date" attributeType="Binary"/>
         <attribute name="id" attributeType="UUID" usesScalarValueType="NO"/>
         <relationship name="nutrientEntries" optional="YES" toMany="YES" deletionRule="Nullify" destinationEntity="NutrientEntry" inverseName="dailyLog" inverseEntity="NutrientEntry"/>
     </entity>
@@ -12,8 +12,4 @@
         <attribute name="portionsRecommended" attributeType="Integer 16" defaultValueString="0" usesScalarValueType="YES"/>
         <relationship name="dailyLog" maxCount="1" deletionRule="Nullify" destinationEntity="DailyLog" inverseName="nutrientEntries" inverseEntity="DailyLog"/>
     </entity>
-    <elements>
-        <element name="DailyLog" positionX="-63" positionY="-18" width="128" height="74"/>
-        <element name="NutrientEntry" positionX="-54" positionY="0" width="128" height="104"/>
-    </elements>
 </model>
